requirements-bot.txt
# TODO: refactor reqirements (cuoff fastAPI ...)
APScheduler==3.6.3
attrs==20.2.0
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptography==3.3.1
decorator==4.4.2
fastapi==0.61.1
h11==0.9.0
httpcore==0.11.1
httpx==0.15.4
idna==2.10
iniconfig==1.0.1
packaging==20.4
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.9.0
pycparser==2.20
pydantic==1.6.1
pyparsing==2.4.7
pytest==6.1.0
pytest-asyncio==0.14.0
python-dotenv==0.14.0
python-telegram-bot==13.1
pytz==2020.4
requests==2.24.0
rfc3986==1.4.0
six==1.15.0
sniffio==1.1.0
SQLAlchemy==1.3.19
SQLAlchemy-Utils==0.36.8
starlette==0.13.6
toml==0.10.1
tornado==6.1
tzlocal==2.1
urllib3==1.25.10
uuid==1.30
uvicorn==0.12.0
